% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_helpers.R
\name{assess_bimodality}
\alias{assess_bimodality}
\title{Assess Bimodality}
\usage{
assess_bimodality(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Bimodality coefficient
}
\description{
Simple test for bimodal distribution which would be expected for
case/control mixture.
}
\keyword{internal}
