% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_helpers.R
\name{assess_kurtosis}
\alias{assess_kurtosis}
\title{Assess Kurtosis}
\usage{
assess_kurtosis(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Kurtosis measure
}
\description{
Simple kurtosis calculation.
}
\keyword{internal}
