% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_helpers.R
\name{assess_skewness}
\alias{assess_skewness}
\title{Assess Skewness}
\usage{
assess_skewness(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Skewness measure
}
\description{
Simple skewness assessment for distribution.
}
\keyword{internal}
