% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_mixture.R
\name{binomial_mixture_m_step}
\alias{binomial_mixture_m_step}
\title{M-step for Binomial Mixture Model}
\usage{
binomial_mixture_m_step(S, C, gamma, regularization = 1e-08)
}
\arguments{
\item{S}{Vector of disease-relevant code counts}

\item{C}{Vector of total code counts}

\item{gamma}{Matrix of posterior probabilities from E-step}

\item{regularization}{Small value for numerical stability}
}
\value{
List with updated parameters p_1, p_0, alpha_0, alpha_1
}
\description{
Updates parameters in the M-step of the EM algorithm for the binomial mixture model.
}
\keyword{internal}
