% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_mixture.R
\name{calculate_binomial_mixture_diagnostics}
\alias{calculate_binomial_mixture_diagnostics}
\title{Calculate Model Diagnostics for Binomial Mixture}
\usage{
calculate_binomial_mixture_diagnostics(
  S,
  C,
  gamma,
  p_1,
  p_0,
  alpha_0,
  alpha_1,
  ll_history
)
}
\arguments{
\item{S}{Vector of disease-relevant code counts}

\item{C}{Vector of total code counts}

\item{gamma}{Posterior probabilities matrix}

\item{p_1}{Estimated success probability for cases}

\item{p_0}{Estimated success probability for controls}

\item{alpha_0}{Estimated intercept}

\item{alpha_1}{Estimated healthcare utilization effect}

\item{ll_history}{Log-likelihood history}
}
\value{
List with diagnostic information
}
\description{
Computes various diagnostic metrics for the fitted binomial mixture model.
}
\keyword{internal}
