% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_helpers.R
\name{create_total_count_query}
\alias{create_total_count_query}
\title{Create Total Count Query}
\usage{
create_total_count_query(table_name, person_ids, date_range, exclude_concepts)
}
\arguments{
\item{table_name}{OMOP table name}

\item{person_ids}{Person ID filter}

\item{date_range}{Date range filter}

\item{exclude_concepts}{Concepts to exclude}
}
\value{
SQL query string
}
\description{
Creates SQL query for extracting total billing code counts from OMOP tables.
}
\keyword{internal}
