% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Check and Establish All of Us Connection}
\usage{
.onAttach(libname, pkgname)
}
\description{
Automatically attempts to establish All of Us database connection when package loads,
but only in All of Us environments.
}
\keyword{internal}
