% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_helpers.R
\name{execute_total_count_query}
\alias{execute_total_count_query}
\title{Execute Total Count Query}
\usage{
execute_total_count_query(query, max_persons = NULL)
}
\arguments{
\item{query}{SQL query string}

\item{max_persons}{Maximum persons to return}
}
\value{
Query results tibble
}
\description{
Executes the total count query against the All of Us database.
}
\keyword{internal}
