% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_phenotypes_impl.R
\name{export_multiple_original_results}
\alias{export_multiple_original_results}
\title{Export Multiple Original Results}
\usage{
export_multiple_original_results(results, output_file)
}
\arguments{
\item{results}{Formatted results from multiple original PheProb}

\item{output_file}{File path for export}
}
\description{
Exports multiple original PheProb results to specified file format.
}
\keyword{internal}
