% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_phenotype_impl.R
\name{export_original_pheprob_results}
\alias{export_original_pheprob_results}
\title{Export Original PheProb Results}
\usage{
export_original_pheprob_results(results, output_file)
}
\arguments{
\item{results}{Formatted results from original PheProb}

\item{output_file}{File path for export}
}
\description{
Exports original PheProb results to specified file format.
}
\keyword{internal}
