% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_helpers.R
\name{extract_domain_from_table}
\alias{extract_domain_from_table}
\title{Extract Domain from Table Name}
\usage{
extract_domain_from_table(table_name)
}
\arguments{
\item{table_name}{OMOP table name}
}
\value{
Domain name
}
\description{
Extracts domain name from OMOP table name.
}
\keyword{internal}
