% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{get_date_column}
\alias{get_date_column}
\title{Get Date Column Name for OMOP Table}
\usage{
get_date_column(table_name)
}
\arguments{
\item{table_name}{Name of the OMOP table}
}
\value{
Character string with column name
}
\description{
Returns the appropriate date column name for each OMOP table.
}
\keyword{internal}
