% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_mixture.R
\name{initialize_binomial_mixture_parameters}
\alias{initialize_binomial_mixture_parameters}
\title{Initialize Parameters for Binomial Mixture Model}
\usage{
initialize_binomial_mixture_parameters(S, C, method = "random")
}
\arguments{
\item{S}{Vector of disease-relevant code counts}

\item{C}{Vector of total code counts}

\item{method}{Initialization method}
}
\value{
List with initial parameters
}
\description{
Provides sensible initial parameter values for the EM algorithm.
}
\keyword{internal}
