% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_helpers.R
\name{is_aou_connected}
\alias{is_aou_connected}
\title{Check All of Us Connection Status}
\usage{
is_aou_connected()
}
\value{
Logical indicating if connected
}
\description{
Checks if the package is connected to All of Us Research Program database.
Only works within All of Us Research Workbench environment.
}
\examples{
\dontrun{
if (is_aou_connected()) {
  # Run analysis
} else {
  connect_aou()
}
}
}
