% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{is_aou_environment}
\alias{is_aou_environment}
\title{Package Initialization}
\usage{
is_aou_environment()
}
\value{
Logical indicating if in All of Us environment
}
\description{
This file contains package initialization functions that run when the package is loaded.
Detect All of Us Environment
}
\details{
Checks if we're running in an All of Us Research Program environment
}
\keyword{internal}
