% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_phenotypes_viz.R
\name{multiple_phenotypes_viz}
\alias{multiple_phenotypes_viz}
\title{Visualization Functions for Multiple Phenotypes}
\description{
This file contains visualization and display functions for multiple phenotype
analysis using the PheProb methodology (Sinnott et al., 2018).
}
