% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_phenotypes_viz.R
\name{plot.pheprob_multiple}
\alias{plot.pheprob_multiple}
\title{Plot Method for Multiple Original PheProb Results}
\usage{
\method{plot}{pheprob_multiple}(x, type = "correlation_matrix", ...)
}
\arguments{
\item{x}{Multiple original PheProb results object}

\item{type}{Type of plot: "correlation_matrix", "probability_distributions",
"comorbidity_network", "convergence_diagnostics"}

\item{...}{Additional arguments passed to plotting functions}
}
\description{
Custom plot method for multiple original PheProb results.
}
