% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_phenotypes_viz.R
\name{plot_correlation_matrix}
\alias{plot_correlation_matrix}
\title{Plot Correlation Matrix}
\usage{
plot_correlation_matrix(results, ...)
}
\arguments{
\item{results}{Multiple phenotype results object}

\item{...}{Additional arguments}
}
\description{
Creates a correlation matrix plot for multiple phenotypes.
}
\keyword{internal}
