% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_phenotypes_viz.R
\name{plot_probability_distributions}
\alias{plot_probability_distributions}
\title{Plot Probability Distributions}
\usage{
plot_probability_distributions(results, ...)
}
\arguments{
\item{results}{Multiple phenotype results object}

\item{...}{Additional arguments}
}
\description{
Creates probability distribution plots for multiple phenotypes.
}
\keyword{internal}
