% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_phenotypes_viz.R
\name{print.pheprob_multiple}
\alias{print.pheprob_multiple}
\title{Print Method for Multiple Original PheProb Results}
\usage{
\method{print}{pheprob_multiple}(x, ...)
}
\arguments{
\item{x}{Multiple original PheProb results object}

\item{...}{Additional arguments (ignored)}
}
\description{
Custom print method for multiple original PheProb results.
}
