% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_phenotype_impl.R
\name{single_phenotype_impl}
\alias{single_phenotype_impl}
\title{PheProb Implementation (Sinnott et al., 2018)}
\description{
This file implements the PheProb methodology using binomial mixture models
as described in Sinnott et al. (2018). This is the main implementation that should be
used for proper probabilistic phenotyping.
}
