% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_phenotypes_impl.R
\name{summarize_convergence}
\alias{summarize_convergence}
\title{Summarize Convergence Across Models}
\usage{
summarize_convergence(phenotype_models)
}
\arguments{
\item{phenotype_models}{List of fitted models}
}
\value{
Convergence summary
}
\description{
Summarizes EM convergence across multiple phenotype models.
}
\keyword{internal}
