% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_phenotypes_viz.R
\name{summary.pheprob_multiple}
\alias{summary.pheprob_multiple}
\title{Summary Method for Multiple Original PheProb Results}
\usage{
\method{summary}{pheprob_multiple}(object, ...)
}
\arguments{
\item{object}{Multiple original PheProb results object}

\item{...}{Additional arguments (ignored)}
}
\description{
Custom summary method for multiple original PheProb results.
}
