% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_mixture.R
\name{validate_binomial_mixture_inputs}
\alias{validate_binomial_mixture_inputs}
\title{Validate Inputs for Binomial Mixture Model}
\usage{
validate_binomial_mixture_inputs(S, C)
}
\arguments{
\item{S}{Vector of disease-relevant code counts}

\item{C}{Vector of total code counts}
}
\description{
Performs input validation for the binomial mixture model.
}
\keyword{internal}
